beaglebone: drop rootfs size restriction (WIP)
This image is to be installed on large-partitioned devices only.

WIP since this needs to be handled differently. My idea is to install
it by default for machines that have been large partitioned from the
start, such as the venus-touch, rpi & easysolar & other nano-pi based
products.

And I'm not sure what to do with the Venus GX yet. Perhaps make two
feeds for it; one for small partitioned versions; and one for large.